Server/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-pandas==0.6.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
greenlet==1.1.0
install==1.3.4
joblib==1.0.1
numpy==1.20.3
pandas==1.2.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
SQLAlchemy==1.4.15
sqlparse==0.4.1
threadpoolctl==2.1.0
